+++ com.ca.lsp.cobol/lsp-service-cobol/src/main/java/com/broadcom/lsp/cdi/module/service/ServiceModule.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Broadcom.
+ * Copyright (c) 2020 Broadcom.
  *
  * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
  *
@@ -16,7 +16,11 @@
 package com.broadcom.lsp.cdi.module.service;
 
 import com.broadcom.lsp.cdi.module.DefaultModule;
+import com.ca.lsp.cobol.model.ConfigurationSettingsStorable;
 import com.ca.lsp.cobol.service.*;
+import com.ca.lsp.cobol.service.delegates.actions.CodeActionProvider;
+import com.ca.lsp.cobol.service.delegates.actions.CodeActions;
+import com.ca.lsp.cobol.service.delegates.actions.FindCopybookCommand;
 import com.ca.lsp.cobol.service.delegates.communications.Communications;
 import com.ca.lsp.cobol.service.delegates.communications.ServerCommunications;
 import com.ca.lsp.cobol.service.delegates.completions.*;
@@ -26,12 +30,17 @@ import com.ca.lsp.cobol.service.delegates.formations.TrimFormation;
 import com.ca.lsp.cobol.service.delegates.references.*;
 import com.ca.lsp.cobol.service.delegates.validations.CobolLanguageEngineFacade;
 import com.ca.lsp.cobol.service.delegates.validations.LanguageEngineFacade;
+import com.ca.lsp.cobol.service.providers.ClientProvider;
+import com.ca.lsp.cobol.service.providers.SettingsProvider;
+import com.ca.lsp.core.cobol.engine.CobolLanguageEngine;
 import com.google.inject.multibindings.Multibinder;
-import com.google.inject.name.Names;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.lsp4j.services.LanguageClient;
 import org.eclipse.lsp4j.services.LanguageServer;
 import org.eclipse.lsp4j.services.TextDocumentService;
+
+import static com.google.inject.multibindings.Multibinder.newSetBinder;
+import static com.google.inject.name.Names.named;
 
 /** This module provides DI bindings for service part. */
 @Slf4j
@@ -46,37 +55,46 @@ public class ServiceModule extends DefaultModule {
     bind(Communications.class).to(ServerCommunications.class);
     bind(TextDocumentService.class).to(MyTextDocumentService.class);
     bind(LanguageClient.class).toProvider(ClientProvider.class);
+    bind(ConfigurationSettingsStorable.class).toProvider(SettingsProvider.class);
+    bind(CobolLanguageEngine.class);
 
     bindFormations();
     bindCompletions();
     bindReferences();
+    bindCodeActions();
   }
 
   private void bindFormations() {
     bind(Formations.class);
-    Multibinder<Formation> formationBinding = Multibinder.newSetBinder(binder(), Formation.class);
+    Multibinder<Formation> formationBinding = newSetBinder(binder(), Formation.class);
     formationBinding.addBinding().to(TrimFormation.class);
   }
 
   private void bindCompletions() {
     bind(Completions.class);
-    Multibinder<Completion> completionBinding =
-        Multibinder.newSetBinder(binder(), Completion.class);
+    Multibinder<Completion> completionBinding = newSetBinder(binder(), Completion.class);
     completionBinding.addBinding().to(VariableCompletion.class);
     completionBinding.addBinding().to(ParagraphCompletion.class);
     completionBinding.addBinding().to(SnippetCompletion.class);
     completionBinding.addBinding().to(KeywordCompletion.class);
     completionBinding.addBinding().to(CopybookCompletion.class);
 
-    bind(CompletionStorage.class).annotatedWith(Names.named("Keywords")).to(Keywords.class);
-    bind(CompletionStorage.class).annotatedWith(Names.named("Snippets")).to(Snippets.class);
+    bind(CompletionStorage.class).annotatedWith(named("Keywords")).to(Keywords.class);
+    bind(CompletionStorage.class).annotatedWith(named("Snippets")).to(Snippets.class);
   }
 
   private void bindReferences() {
     bind(Occurrences.class).to(SemanticElementOccurrences.class);
     Multibinder<SemanticLocations> referenceBinding =
-        Multibinder.newSetBinder(binder(), SemanticLocations.class);
+        newSetBinder(binder(), SemanticLocations.class);
     referenceBinding.addBinding().to(VariableLocations.class);
     referenceBinding.addBinding().to(ParagraphLocations.class);
+  }
+
+  private void bindCodeActions() {
+    bind(CodeActions.class);
+    Multibinder<CodeActionProvider> codeActionBinding =
+        newSetBinder(binder(), CodeActionProvider.class);
+    codeActionBinding.addBinding().to(FindCopybookCommand.class);
   }
 }